requirements.txt
bleach>=3.0.2
Django>=2.1.4
django-el-pagination>=3.2.4
django-hitcount>=1.3.0
django-markdownx>=2.0.28
django-simple-captcha>=0.5.9
django-taggit>=0.23.0
django-widget-tweaks>=1.4.3
psycopg2>=2.7.6.1
Pygments>=2.3.0
wagtail>=2.4
wagtailmenus>=2.12
